requirementsnew.txt
tensorflow 
numpy 
keras==3.0.0 
Flask
Werkzeug
opencv-python
Pillow
